feat: update assigned applications card, add pending standard applications card
